solved modules task
